feat: 핫 유저 구현
