+++ trAIner/app/api/v1/__init__.py
@@ -6,4 +6,7 @@ from flask import Blueprint
 api_v1 = Blueprint('api_v1', __name__)
 
 
-from . import calculator, info, admin, problem, user, hot_user, cold_user
+from . import (
+    calculator, info, admin,
+    problem, user
+)
